old/requirements.txt
Flask>=2.2
Flask-WTF>=1.1.1
WTForms>=3.0.1
Flask-Login>=0.6.2
Flask-Bcrypt>=1.0.1
Flask-Migrate>=4.0.4
Flask-SQLAlchemy>=3.0.3
python-dotenv>=1.0.0
requests>=2.31.0
gunicorn>=20.1.0
email-validator>=1.3.1
numpy>=1.25.2
opencv-python-headless>=4.7.0
pyzbar>=0.1.8
reportlab>=3.6.14
arabic-reshaper==3.0.0
python-bidi==0.4.2
